Extract getPath from common
